Revert "drm/atomic: Boost CPU and DDR bus when committing a new frame"
While this is a resonable solution to handling interactions which extend
beyond the critical latency window covered by input boosting when
userspace boosts are blocked in-kernel, it also carries with it the
inherent drawback of being unable to differentiate between user
interactions and graphically intensive activities such as video
playback. This creates scenarios in which minimum frequencies are
needlessly locked to boosted values for indefinite periods of time.

This reverts commit 7bf6cc8 in preparation for an alternative approach
to handling long-term interactions while eliminating the potential for
extraneous boostlocks.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>