sched: Disable SCHEDTUNE_BOOST_HOLD_ALL
The decision to enable this was excessive to say the least, especially
given the elaborate nature of our current boosting framework. I have
considered doing away with this feature entirely, but at the very least
let's restore the default behavior of applying boost holds to the RT
task group only while additional testing is carried out.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>