sched/boost: Boost CPU and DDR bus when sched_boost is triggered
We have the benefit of perfd writing a nonzero value to
/proc/sys/kernel/sched_boost during extended interactions such as
scrolling/flings, as well as top-app migration. Despite the subsequent
userspace boosts being blocked in-kernel, we can integrate this node
with cpu_input_boost and repurpose it as a boost kick trigger to handle
long-term user interaction which extends beyond the critical latency
period covered by input boosting without the need to indiscriminately
boost on frame committal.

This ammends commit cce32d2 ("sched/boost: Perform SchedTune boosting
when sched_boost is triggered").

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>