Revert "kbuild: Support Polyhedral Loop Optimizations for LLVM"
Most if not all custom ROMs are being compiled with AOSP clang by
default, which does not natively support Polly. This only serves to
create needless headaches and confusion for inline builders as
CONFIG_POLLY_CLANG would need to be disabled in order to build under
most circumstances. In this case, we can eliminate the extra step and
just remove this option entirely.

This reverts commit 342527d8a5b1b9fab5a00ad3ab85e9bdc462ca20.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>